core/requirements.txt
beautifulsoup4==4.14.3
python-dotenv==1.2.2
Requests==2.32.5
urllib3==2.6.3
